+++ AnxietyManager/app/src/main/AndroidManifest.xml
@@ -12,26 +12,26 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".CreateAccountActivity">
+        <activity android:name=".activities.CreateAccountActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SensedActivity"></activity>
-        <activity android:name=".HomeActivity"></activity>
-        <activity android:name=".WhatsUpActivity"
+        <activity android:name=".activities.SensedActivity"></activity>
+        <activity android:name=".activities.HomeActivity"></activity>
+        <activity android:name=".activities.WhatsUpActivity"
             android:label=" "
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"
             android:windowSoftInputMode="stateHidden" ></activity>
-        <activity android:name=".ThoughtsActivity"
+        <activity android:name=".activities.ThoughtsActivity"
             android:windowSoftInputMode="stateHidden" ></activity>
-        <activity android:name=".PhysicalActivity"></activity>
-        <activity android:name=".MoodActivity"></activity>
-        <activity android:name=".RateTwoMoodActivity"></activity>
-        <activity android:name=".RateOneMoodActivity"></activity>
-        <activity android:name=".ReactActivity"></activity>
+        <activity android:name=".activities.PhysicalActivity"></activity>
+        <activity android:name=".activities.MoodActivity"></activity>
+        <activity android:name=".activities.RateTwoMoodActivity"></activity>
+        <activity android:name=".activities.RateOneMoodActivity"></activity>
+        <activity android:name=".activities.ReactActivity"></activity>
 
     </application>
