+++ AnxietyManager/app/src/main/java/com/example/katiefitzgerald/anxietymanager/activities/CreateAccountActivity.java
@@ -1,10 +1,11 @@
-package com.example.katiefitzgerald.anxietymanager;
+package com.example.katiefitzgerald.anxietymanager.activities;
 
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.os.Parcelable;
 import android.support.v4.app.ActivityCompat;
+import android.support.v7.app.AppCompatActivity;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
@@ -12,12 +13,14 @@ import android.widget.CheckBox;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.example.katiefitzgerald.anxietymanager.R;
+import com.example.katiefitzgerald.anxietymanager.model.UserDao;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 import pub.devrel.easypermissions.EasyPermissions;
 
-public class CreateAccountActivity extends SensedActivity {
+public class CreateAccountActivity extends AppCompatActivity {
 
     EditText enterEmail;
     EditText enterPassword;
