+++ AnxietyManager/app/src/main/java/com/example/katiefitzgerald/anxietymanager/activities/PhysicalActivity.java
@@ -1,4 +1,4 @@
-package com.example.katiefitzgerald.anxietymanager;
+package com.example.katiefitzgerald.anxietymanager.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.ImageView;
+
+import com.example.katiefitzgerald.anxietymanager.R;
 
 
 /**
