+++ AnxietyManager/app/src/main/java/com/example/katiefitzgerald/anxietymanager/activities/RateTwoMoodActivity.java
@@ -1,4 +1,4 @@
-package com.example.katiefitzgerald.anxietymanager;
+package com.example.katiefitzgerald.anxietymanager.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,6 +9,8 @@ import android.widget.ImageView;
 import android.widget.SeekBar;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.katiefitzgerald.anxietymanager.R;
 
 /**
  * Created by Katie Fitzgerald on 30/01/2018.
