+++ AnxietyManager/app/src/main/java/com/example/katiefitzgerald/anxietymanager/activities/ThoughtsActivity.java
@@ -1,4 +1,4 @@
-package com.example.katiefitzgerald.anxietymanager;
+package com.example.katiefitzgerald.anxietymanager.activities;
 
 import android.content.Intent;
 import android.database.Cursor;
@@ -16,6 +16,11 @@ import android.widget.ImageButton;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.katiefitzgerald.anxietymanager.sql.DatabaseManager;
+import com.example.katiefitzgerald.anxietymanager.model.QuestionnaireDao;
+import com.example.katiefitzgerald.anxietymanager.R;
+import com.example.katiefitzgerald.anxietymanager.adapters.ThoughtsAdapter;
 
 import java.sql.SQLException;
