+++ AnxietyManager/app/src/main/java/com/example/katiefitzgerald/anxietymanager/adapters/SensedAnxietyAdapter.java
@@ -1,4 +1,4 @@
-package com.example.katiefitzgerald.anxietymanager;
+package com.example.katiefitzgerald.anxietymanager.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -6,6 +6,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
+
+import com.example.katiefitzgerald.anxietymanager.model.AnxietyEpisode;
+import com.example.katiefitzgerald.anxietymanager.R;
 
 import java.util.ArrayList;
