+++ AnxietyManager/app/src/main/java/com/example/katiefitzgerald/anxietymanager/model/AnxietyEpisode.java
@@ -1,4 +1,4 @@
-package com.example.katiefitzgerald.anxietymanager;
+package com.example.katiefitzgerald.anxietymanager.model;
 
 /**
  * Created by Katie Fitzgerald on 20/11/2017.
